Rearrange code to be more readable.  Remove some dead code.  Refactor a little to remove duplicate code.
